old-tests/submission/Microsoft/video/video_014.htm

<!DOCTYPE HTML>
<html>
<head>
  <title>HTML5 Media Elements: 'buffered' returns a TimeRanges object.</title>
  <meta content="text/html; charset=UTF-8" http-equiv="Content-Type">
  <link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
  <link rel="help" href="http://www.w3.org/TR/html5/video.html#dom-media-buffered" />
  <meta name="assert" content="'buffered' returns a TimeRanges object." />
</head>
<body>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>

    <div id=log></div>

    <video id="video0">Your browser does not support media elements.</video>
    <script type="text/javascript">
        test(function() {assert_true(document.getElementById("video0").buffered instanceof TimeRanges)}, "'buffered' returns a TimeRanges object.");
    </script>
</div>
</body>
</html>
